+++ stwo_cairo_prover/crates/cairo-air/src/poseidon/air.rs
@@ -30,7 +30,7 @@ impl PoseidonContextClaim {
             .unwrap_or_default()
     }
 
-    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u32>) {
+    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u64>) {
         if let Some(claim) = &self.claim {
             claim.accumulate_relation_uses(relation_counts);
         }
@@ -67,7 +67,7 @@ impl Claim {
         TreeVec::concat_cols(log_sizes)
     }
 
-    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u32>) {
+    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u64>) {
         let Self {
             poseidon_3_partial_rounds_chain,
             poseidon_full_round_chain,
